feat: :sparkles: Remove user, fixed double states, removed unnecessary code

+++ src/userForm/store.ts
@@ -26,9 +26,14 @@ export const useUserStore = create<UserStore>((set, get) => ({
 
   addUser: (user) => {
     set((state) => {
-      const updatedUsers = [...state.users, user];
-      localStorage.setItem("users", JSON.stringify(updatedUsers));
-      return { users: updatedUsers };
+      const userExists =
+        state.users.includes(user) || state.bookmarked === user;
+      if (!userExists) {
+        const updatedUsers = [...state.users, user];
+        localStorage.setItem("users", JSON.stringify(updatedUsers));
+        return { users: updatedUsers };
+      }
+      return state;
     });
   },
